02: refactored manager components
