+++ src/components/CrudDialog.jsx
@@ -1,5 +1,3 @@
-// src/CrudDialog.jsx
-import React from 'react';
 import { Dialog, DialogTitle, DialogContent, DialogActions, Button, TextField } from '@mui/material';
 
 export default function CrudDialog({
