+++ src/components/CrudTable.jsx
@@ -1,5 +1,3 @@
-
-
 import React from 'react';
 import {
   Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper, TablePagination, Button
